+++ src/server/views_by_date.rs
@@ -91,8 +91,8 @@ sql_function! {
 }
 
 mod filter {
-    use diesel::sql_types::{Nullable, Timestamp};
     use diesel::sql_function;
+    use diesel::sql_types::{Nullable, Timestamp};
 
     sql_function! {
         fn year_of_timestamp(date: Nullable<Timestamp>) -> Nullable<SmallInt>