+++ CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on
 
 * Improved item (image) sizing css.  I hope it works in safari.
 * Updated to `diesel` 2.1.0, `diesel-async` 0.3.2, and `ructe` 0.17.0.
+* Uses std LazyLock (require rust 1.80) instead of `lazy_regex`.
 * Some more known kinds of places.